Fix solution to account for proper elimination
Previously, the solution relied on the fact that the person who was last passed the pumpkin would be eliminated. This is in fact incorrect, as the person who last passed the pumpkin (the last person to say a number) was actually the person to be eliminated. This resulted in a classic off by one error, which is resolved in this commit.